+++ src/increment/api/controllers/increment_v2.py
@@ -30,19 +30,3 @@ class IncrementController(APIController):
         count = await service.get_count()
 
         return count
-
-
-class IncrementControllerV2(APIController):
-    @classmethod
-    def version(cls) -> str:
-        return "v2"
-
-    @classmethod
-    def class_name(cls) -> str:
-        return "increment"
-
-    @post()
-    async def increment(self):
-        """
-        Faster version of increment, v2
-        """
